copied settings from strawbees test

+++ .platform/hooks/postdeploy/10_setup_ssl_certificates.sh
@@ -2,10 +2,10 @@
 
 # Install certbot
 echo "LOG: Installing certbot"
-sudo wget -r --no-parent -A 'epel-release-*.rpm' http://dl.fedoraproject.org/pub/epel/7/x86_64/Packages/e/
-sudo rpm -Uvh dl.fedoraproject.org/pub/epel/7/x86_64/Packages/e/epel-release-*.rpm
-sudo yum-config-manager --enable epel*  > /dev/null
-sudo yum -y install certbot python2-certbot-nginx
+wget -r --no-parent -A 'epel-release-*.rpm' http://dl.fedoraproject.org/pub/epel/7/x86_64/Packages/e/
+rpm -Uvh dl.fedoraproject.org/pub/epel/7/x86_64/Packages/e/epel-release-*.rpm
+yum-config-manager --enable epel*  > /dev/null
+yum -y install certbot python2-certbot-nginx
 
 # Generate certificates
 echo "LOG: Generating certificates"
@@ -14,19 +14,26 @@ for domain in $(echo ${LETSENCRYPT_DOMAINS} | sed "s/,/ /g")
 do
   certbot_command="$certbot_command -d $domain"
 done
-eval $certbot_command
 
-# Link certificates
-echo "LOG: linking certificates"
-rm -rf /etc/letsencrypt/live/ebcert
-domain="$( cut -d ',' -f 1 <<< "${LETSENCRYPT_DOMAINS}" )";
-if [ -d /etc/letsencrypt/live ]; then
-  domain_folder_name="$(ls /etc/letsencrypt/live | sort -n | grep $domain | head -1)";
-  if [ -d /etc/letsencrypt/live/${domain_folder_name} ]; then
-	ln -sfn /etc/letsencrypt/live/${domain_folder_name} /etc/letsencrypt/live/ebcert
-  fi
+if eval $certbot_command; then
+    # Link certificates
+    echo "LOG: linking certificates"
+    rm -rf /etc/letsencrypt/live/ebcert
+    domain="$( cut -d ',' -f 1 <<< "${LETSENCRYPT_DOMAINS}" )";
+    if [ -d /etc/letsencrypt/live ]; then
+      domain_folder_name="$(ls /etc/letsencrypt/live | sort -n | grep $domain | head -1)";
+      if [ -d /etc/letsencrypt/live/${domain_folder_name} ]; then
+    	ln -sfn /etc/letsencrypt/live/${domain_folder_name} /etc/letsencrypt/live/ebcert
+      fi
+    fi
+    # Setup cron job for certificate renewal
+    rm -f /etc/cron.d/certificate_renew /etc/cron.d/certificate_renew.bak
+    echo "17 2,14 * * * root date >> /var/log/cron.log && certbot renew --post-hook \"systemctl restart nginx\" >> /var/log/cron.log" > /etc/cron.d/renew_certificate
+    chmod 644 /etc/cron.d/certificate_renew
+else
+    echo "LOG: There was an error generating the certificates. Https may not work correctly."
 fi
 
 # Reload nginx
-echo "LOG: Reloading nginx"
-service nginx reload
+echo "LOG: Restarting nginx"
+systemctl restart nginx
